requirements.txt
anthropic==0.42.0
huggingface_hub==0.23.2
jsonlines==3.1.0
numpy==1.24.4
orjson==3.10.13
pandas==2.0.3
Pillow==11.1.0
tqdm==4.65.0
trio==0.27.0+dev
